backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
python-dotenv==1.1.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
